% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em.R
\name{em}
\alias{em}
\title{em}
\usage{
em(
  dat,
  subs,
  mus,
  sigma,
  likfun,
  bic = F,
  ndata = NULL,
  emtol = 1e-05,
  maxiter = 100L,
  lower = -Inf,
  upper = Inf,
  parallel = F,
  return_ind_df = F,
  verbose = T,
  ...
)
}
\arguments{
\item{dat}{data.frame; data}

\item{subs}{string vector; subject ids}

\item{mus}{vector; starting mean parameters}

\item{sigma}{vector; starting sd in parameter}

\item{likfun}{function; the objective function}

\item{bic}{logical; if TRUE, calculate the group-level bic after fitting model}

\item{ndata}{integer; number of observations. only needed to calculate bic}

\item{emtol}{numeric; tolerance for convergence}

\item{maxiter}{integer; maximum number of iterations. Will stop if maxiter reached without convergence}

\item{lower}{numeric vector; lower bounds for parameters}

\item{upper}{numeric vector; upper bounds for parameters}

\item{parallel}{logical; if TRUE, will optimize subjects in parallel (within an iteration)}

\item{return_ind_df}{logical; if TRUE, will only return a data frame of individual results (parameters and likelihood)}

\item{verbose}{logical; if TRUE, will print progress at each iteration}
}
\value{
if return_ind_df = FALSE, returns a list containing the following fields:
\item{mus}{vector of mean parameters}
\item{sigma}{vector of standard deviation of each parameter}
\item{x}{npar x nsubjects matrix with individual parameters}
\item{l}{vector of likelihood for each subject}
\item{h}{npar x npar x nsubjects array. each slice contains the hessian matrix for each subject at that subjects individual parameters}

if return_ind_df = TRUE, returns a data frame with the following columns:
\item{Subject}{Subject ID}
\item{Parameters}{a separate column for each parameter}
\item{lik}{The likelihood for the individual subject}
\item{bic}{The group-level bic for the model (only if)}
}
\description{
Run expectation maximization algorithm to fit parameters to each subject of an experiment
}
