% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ind_df.R
\name{get_ind_df}
\alias{get_ind_df}
\title{get_ind_df}
\usage{
get_ind_df(em_fit)
}
\arguments{
\item{em_fit}{list; output from GaussExMax::em}
}
\value{
returns a data frame with the following columns:
\item{Subject}{Subject ID}
\item{Parameters}{a separate column for each parameter}
\item{lik}{The likelihood for the individual subject}
\item{bic}{The group-level bic for the model (only if)}
}
\description{
Get data frame of individual parameters from EM fit
}
