% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibic.R
\name{ibic}
\alias{ibic}
\title{ibic}
\usage{
ibic(em_fit, ndata)
}
\arguments{
\item{em_fit}{list; output from GaussExMax::em}
}
\value{
numeric; integrated bayes information criterion
}
\description{
Get integrated bayes information criterion from em fit
}
