% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lml.R
\name{lml}
\alias{lml}
\title{lml}
\usage{
lml(em_fit)
}
\arguments{
\item{em_fit}{list; output from GaussExMax::em}
}
\value{
numeric; the log marginal likelihood across subjects
}
\description{
Get log marginal likelihood from an em fit
}
